third_party/blink/web_tests/external/wpt/fetch/metadata/generated/element-embed.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] sec-fetch-mode
  assert_array_equals: expected property 0 to be "no-cors" but got "navigate" (expected array ["no-cors"] got ["navigate"])
[FAIL] sec-fetch-storage-access - Cross-site
  assert_array_equals: expected property 0 to be "none" but got "active" (expected array ["none"] got ["active"])
Harness: the test ran to completion.
